Remove id from label,input and change styling of the read status checkbox
Why i remove the id? Because of the lack of uniqueness on cards
and this caused a problem that if you click the label of any element
it will always refer to first element. Because of that i changed the
styling of the checkbox to make it bigger and better in terms of visuals.